feat: build endpoint for api (#817)
* feat: build protos

Based on deployment protos. Not final, will need to be updates, but want
them here so I can wire up the API.

* feat: set up builds server in API

Copied from deployments server and did some search/replace. Definitely
not working yet, but need this to wire up the gRPC API.

* feat: add builds server to api

* feat: add build model

* feat: remove request methods we're not using

May add this stuff back after basic create and read methods are working.

* refactor: not sure why we rename the connect

* feat: just log to test request handlers are working

* feat: need to convert gorm to proto

Looks like each gorm type needs a method to convert it to a proto. Feels
wrong to have to write all this manually. At the very least, we can get
rid of the converters and make this part of the model contract.

* refactor: move build client init to func ops

Not clear why we're initializing the builds server clients in the
server, rather than sticking with the functional options pattern we're
using for everything else.

I have some doubts about the usage of functional ops as well, but that
seems like more refactoring than I want to take on in this change.

* fix: forgot to actually init the build server

* feat: new workflow methods

Instead of using CRUD nomenclature, borrowing from Temporal's SDK to
expose methods that make more sense for workflows:

- Start: to start a workflow
- Cancel: to cancel a workflow
- Query: to search for workflows

Not clear that we actually need a GET, at least right now, but can
easily be added.

* feat: API calls Temporal

We return a temporal error, since the workflow doesn't exist yet, but
once it's added this should work.

Might take a second pass at the handler code itself, maybe put some of
this into a method on the build model. It's a fair amount of logic to
record the request in the DB, start the workflow, get the workflow back
and then combine the data.

* feat: change "commit_hash" to "git_ref"

* WIP: defining the Build temporal workflow

* feat: add builds table

* feat: add db client to builds server

* fix: build need a git ref but not an install

* refactor: use shordIDs for build protos and model

Signed-off-by: Maria Paktiti <maria@powertools.dev>

* fix: needed to rename my migration after rebase

* fix: remove broken test, will add new ones later

Want to merge this asap since other work is happening on it. Will have a
few more PRs later to add tests, and do some refactoring.

* feat: set IDs when creating build record

Uses new shortid lib.

---------

Signed-off-by: Maria Paktiti <maria@powertools.dev>
Co-authored-by: Peter Lyons <pete@powertools.dev>
Co-authored-by: nuonbot <128654611+nuonbot@users.noreply.github.com>
Co-authored-by: Maria Paktiti <maria@powertools.dev>