refactor: builds server uses shared temporal client (#898)
We don't need to create a new temporal client instance for each server.
Updating the builds server to use the client already being created in
the cmd.

This works, but it started logging that the temporal client wasn't given
a logger, even though I'm giving it the app logger in the cmd. Need to
dig into that.