fix: add in the github info we were missing (#906)
* fix: add in the github info we were missing

This was previously being handled in the deployments service. This
probably isn't what we want to do long-term, but for now we just need
this info appended to each plan request.

* fix: copy deployments service code we need for builds

My first attempt at this didn't work. I thought I was simply updating
the build config with github info. Turns out the build config was nil,
and needed to be constructed.

I stepped backwards through the service code path, until I understood
exactly what it was doing, then copied over the minimum code I needed to
get it working.

This needs some cleanup, but it has least has all the business logic.

* fix: update build job tests with extra params