fix: assume role + executor plugin lookup bugs (#1043)
* fix: ensure that session duration is not too large

We recently started setting the assume role session duration. AWS states
that the duration is max of an hour, when chaining assume roles (ie:
assuming a role, while assuming a role).

We had hard coded this as `time.Hour`, but due to rounding issues, that
is actually larger than 3600 seconds.

* fix: ecr public needs to use us-east-1
changes. Lines starting

* fix test