fix: fixes for sandboxes (#1259)
* fix: ignore terraform config files

* fix: pin eks module version

Pin the eks module minor version. New sandboxes seem to be failing due
to EKS nodes not joinging the cluster. I haven't been able to test this
locally, but my theory is that going from 19.0 to 19.15 may be the root
cause of this.

Obviously not a long-term solution, but hopefully this can unblock us.

* fix: don't check in tf plan files

* fix: roll back to v0.11.1

Looks like dependabot updates made to the legacy sandboxes repo since we
tagged v0.11.1 caused a breaking change somewhere. Rolling back to that
tag seems to have fixed the EKS nodes not joining the node group.

Will re-apply my changes on top of this.

* feat: give ODRs all permissions

Give ODRs all permissions so they can run terraform. This is a temporary
solution until we've figured out how to implement properly scoped roles
based on job types.

---------

Co-authored-by: nuonbot <128654611+nuonbot@users.noreply.github.com>