fix: ja/1770-odr-k8s-permissions (#1912)
* Revert "fix: referencing wrong name for waypoint helm chart"

This reverts commit 26a078fa33c6026f29ec76bc0af806bcded8c2e5.

Still got the same error in CI after this change. After some local
testing, it looks like helm generates the filename it wants, and I can't
change it. Not sure exactly how to proceed with this.

Also not clear if we need the OCI artifact for this, because it's also
pushing to S3.

* Revert "fix: give odr admin permisions over entire k8s cluster (#1910)"

This reverts commit f70a529008cbccdda3c202a17226f79e4c20d916.

We needed to update the clusterrole, not the namespaces roles. This
chart will only create these edit roles and bindings if you configure it
with namespaces to create them for. We don't want any namespaced roles,
because we want the intall runner to admin the entire cluster.

* fix: give ODRs cluster admin permissions

we don't need this for the build runner, just the install runner. But,
this chart doesn't know about that distinction, and it probably doesn't
matter if we give the build runner admin permissions for now.

Now that the waypoint helm chart is forked, we could update it to
differentiate the build and install runners, but that's outside the
scope of what we need for this week.