fix: enable liveness check for api-gateway (#1982)
The liveness check was disabled when the api-gateway was copied into the
monorepo, so I'm not sure why this off in the first place. But, without
a liveness check, the service won't restart when it's unhealthy.

This won't fix the underlying issue of the api-gateway exceeding it's
cpu limit in the first place. But, at least unhealthy pods should be
replaced and we won't be impacted while trying to use the API.