feat: migrate app resource to new api (#2130)
* feat: add app resource tests

A couple of weird errors here, might just move to the new API first to
get around it.

First, I got an error from the provider that the org_id was in the
resource path, but no the resource schema. We don't want the org_id as
an attribute anymore, I think, now that we're using it as part of auth.
It would be like all AWS resources having the AWS account ID as an
attribute.

But, removing org_id from the path results in an API error. The org ID
doesn't get set correctly on the request, I guess, and it says it's not
able to upsert the org. I assume this won't be an issue with the new
API. I was hoping to get the test working first and then move it over,
but that may not be worth the effort.

* feat: add new api client

* feat: app resource uses new api

* chore: disable artifact "latest" releases

We'll keep these off until the clients have been updated and are stable.