feat: Ja/issue 2073 new api sdk (#2170)
* feat: add rest api cient to data sources

* feat: app data source uses new api

* chore: define API statuses

These should come from the SDK. Defining them here for now, until the
SDK is updated.

* feat: remove build resource

We're not using this currently, and doesn't fit into Terraform's data
model.

Builds are not mutable, and can't be updated or deleted, therefore they
don't fit into Terraform's data model. It's possible Terraform may add
support for resources like this in the future, but for now, it's
probably best to cut scope here.

* feat: remove the deploy resource

Same as with builds, this doesn't fit into Terraform's data model.
Removing it for now.

* fixup: remove build resource

* fixup: remove deploy resource

* feat: install resource uses new api

* feat: remove org data source

This isn't needed anymore.

* feat: repos data source uses new api

* feat: repo data source uses new api

* feat: remove repos data source

We're not using this, and it's not clear there's a use-case for it. It's
more likely users will just want the single repo data source.

* feat: install data source uses new api

* feat: container image resource uses new api

* feat: docker build resource uses new api

* feat: helm chart resource uses new api

* feat: terraform resource uses new api

* refator: remove leftover graphql client references