feat: remove release strategy attribute (#2280)
Releases no longer have a "strategy" attribute. Instead of having
"parallel" and "sync_with_delay", all releases work the same.

Installs per step must be set. If set to a positive number, that many
installs will be deployed to at a time. If set to 0, all installs will
be deployed to at once. If set to a negative or any other number, the
release creation request will fail.