feat: Ja/cli env vars (#2288)
* feat: new config init to support env vars and config files

Just merging the function definition for now. Need to merge in a few
other changes and I need some more time to test this before enabling it.

Using this config init the CLI will read config values from flags,
env vars, and a config file, in that order.

The env vars and config file values will be determined by the cobra
flags set on each command.

* feat: don't require API_URL and APP_ID env vars

All commands that don't require an app ID will work without one set.
Commands that do require an app ID will fail. The resulting error
message is clear enough, but it's not a great UX. Cobra would handle
this better if we weren't going around it's flags and using env vars.

IMO, the UX is good enough for now, and will be improved with a later
commit that will add flags as well as support env vars.

---------

Co-authored-by: nuonbot <128654611+nuonbot@users.noreply.github.com>