fix: CLI config wasn't binding at root. (#2313)
Turns out that cobra's PersistentPreRunE attribute is only inherited by
direct child commands. So, it was being inherited by the children of the
root command, but not by the grandchildren.

For now, I just copy/pasted the attribute to all subcommands. I think we
can avoid having to do this with some refactoring, but I'd like to get
these UX changes released first.