refactor: pass context via context (#2318)
We were passing a context.Context around via closure, which made it
impossible to move the business logic in the command handlers into
their own packages. cobra allows you to provide a context when
initializing it, that will be passed to all subcommands, essentially
acting like a request context would in an API router.