refactor: move all CLI business logic out of cmd (#2322)
This applies the same changes to all commands that were previously
applied to installs, and then does some cleanup.

I also renamed the `context` command to `orgs current`. Creating a
package for each command meant we would have `/internal/context`, which
was obviously inconvenient. Since the command is essentially showing you
the org you're authenticated with, I decided to create an `orgs`
namespace, and then add a `current` command under that.

Once this is merged, there a few more changes coming:

- Add a command to view builds.
- Once we can view builds, we can implement creating releases.
- We may want to redesign how builds and releases are nested under
  components. I think that's going to get awkward as we add more
  commands.