feat: cli builds commands (#2324)
I tried adding these under the components command, but it was awkward to
implement and use, so I created a top-level builds command.

I noticed that we need the component ID to get a build, even when we
already have the build ID. Are build IDs not unique? Feels wrong.

I also realized that the "register" functions used for the commands are
actually constructors, so I'll be renaming them. This also made me
realize I could refactor the root command to further simplify the code
and clarify some boundaries in the code. I'll have a follow-up PR for
that soon.