refactor: decouple cli command constructors from init code (#2325)
Previously, the Execute function was acting as both init code for the
executable -- what you expect in the cmd directory -- as well as a
constructor the root command. This PR pulls those two roles apart, and
treats all the commands as modular components with their own
constructors, that are then composed together by the Execute function.

It should be possible now to pull the commands themselves out of cmd,
maybe into something like `/internal/commands`. I'm not sure about that
yet, though, so I'll leave them where they are for now.