refactor: clean up the cli flags a bit (#2345)
Goal is to maintain consistency for important flags that should
be the same everywhere. Mainly this means the ID flags, so `-a` is
always `--app-id`, `-r` is always `--release-id`. For some of the
sub-commands this is challenging.

It might be better to just not have short forms for flags specific to
one sub-command.

We may also want to look into upgrading our version of cobra.