feat: basic views for cli builds commands (#2361)
Setting up basic views for the CLI commands, startin with builds. I
wanted to iterate a bit using builds before using these for everything.

The views themselves still aren't as polished as I would like, but the
basic patterns for using them is established, and in a place where we
should be able to iterate on them pretty easily.