fix: stop concurrent deploys (#2472)
We have a problem with deploys happening concurrently and interfering
with each other. Setting `concurrency` on the service_deploy workflow
should prevent that.

If this workflow is running, subsequent runs will wait in a "pending"
status until it's complete.

If any runs are already pending when a new run is triggered, they will
be cancelled. The most recent run will always "win".