fix: release steps will now be created (#2474)
`generics.SliceToGroups` was appending the install IDs for a release to
each row slice, but not appending the row back to the column slice.
Since Golang is pass by value, and you get a new slice from each call to
`append`, the function was always returning an empty array.

At least, that's what I was seeing. This change references the row slice
within the column slice to avoid this issue.