refactor: move config logic into it's own package (#2513)
I wanted to do this when first setting up the CLI config, but it didn't
seem necessary then. Now I want to reuse some of this for the TF
provider, so decoupling it from CLI's cmd code makes sense.

In order to truly share the config logic, we'd need to put the config
library it it's repo and publish that. I'm not sure we need to go that
far just yet. For now I should be able to copy the code I need, and
validate if it really will be identical.