fix: can list all components (#2772)
`nuon components list` without an app ID would return the unexpected
error message, even though the app ID is optional. This was happening
because we were trying to use the value of the `--app` flag before
checking if it was empty.

Also cleaned up some error printing duplicate code in the ui library.