fix: Ja/ecs demo fixes (#4224)
fix(bins/cli): pad to hide pterm line overwrite bug

Our previous work-around was never meant to be permananent, but it looks
like the race condition bug is still not fixed in pterm.

https://github.com/pterm/pterm/pull/447

This is a slightly better solution. Instead of printing an arbitrary
amount of spaces, we print enough to overwrite the previously printed
text, which is all we really needed in the first place.