fix(infra/datadog): fix synthetic location name (#4503)
* fix(infra/datadog): fix synthetic location name

* fix(infra/datadog): stop updating the monitors on each apply

Every time we apply this TF, it keeps adding two options:

- grouby_simple_monitor: false
- new_host_delay: 300

These weren't included in the exported JSON. It looks like the Datadog
provider includes these on it's own in the monitor resource.

This isn't technically a bug. It's not breaking anything, but it does
trigger an apply even though nothing has changed, which is noisy and
annoying. Hopefully just adding them to the JSON will prevent this.

* refactor(infra/datadog): correct indentation in yaml file