feat(infra/datadog): adjust org health monitor (#4684)
Adjust the org health check monitor to alert if a health fails 3 times
within 10 minutes. It looks like org health checks sometimes fail once or twice, then
recover. This could be due to:

- k8s rotating pods
- canary runs quickly taking up resources

Regardless of the reason, the org appears to recover quickly. I think we
only need to be alerted if an org becomes unhealthy and is unable to
recover.