feat(bins/cli): add "--all" flag to validate command (#4964)
* feat(bins/cli): add "--all" flag to validate command

Kind of tedious validating each app config individually. Saw this on the
sync command and copied it.

* feat(bins/cli): default to false for "--all" flags

The error message in the sync command made it look like the user is
suppoed to explicitly set the "--all" flag, which makes sense. By
defaulting to "true", you can accidentally sync all config files in the
directory if you forget to pass in a file name.