feat(services/ctl-api): set install status (#5266)
Need this so clients can configure UI. For example: disable the
Provision/Reprovision button in the installer.

Note that the status still isn't 100% accurate. The current logic can't
determine if the install is provisioning or deprovisioning. Simply
sending "provisioning" is good enough for now, but we may want to update
this later.