feat: Ja/fix run local code generation (#5559)
* fix(bins/nuonctl): ignore generate files when running services locally

When running `nuonctl services run-local --name ctl-api`, it was stuck
in a loop, restarting over and over because it detected a change to the
file `services/ctl-api/internal/app/components/worker/activities/get_component_gen.go`.

We run `go generate` as part of `run-local`. It filters out certain
directories, but this new file was in a regular directory, alongside
manually-written code. This commit filters out files that end with
`_gen.go`. Not if that's the best solution but it prevents `run-local`
from getting stuck in a loop.

* feat(services/ctl-api): update service config to use regex for ignore

* feat(services): update all services using ignore to use regex

While testing this change, I found that nuonctl gets stuck in a loop
running dashboard-ui. I checked out main and the same thing was
happening there. It looks like this was happening before my change, and
we didn't notice because we're not using nuonctl to run the JS projects.

Pushing this to keep the app configs in sync. Will investigate the
ingore patterns separately.

---------

Co-authored-by: nuonbot <bot@nuon.co>