feat: ctl-api/otel: small featch logs endpoint w/ filters  (#6036)
* fix: ensure we pass time.Time to timestamps fields

to avoid any potential for losing nanoseconds

* feat: ctl-api/otel: read logs endpoint w/ query params

job_id and job_execution_id accepted as query params

* fix: ctl-api/otel: ensure we pass a time.Time to all time.Time fields

otherwise they are not NULL, they are ZERO! becuse of golang/gorm struct
default values. this is highly problematic because the timestamp_date is
used for TTLs and the zero is a date back in 1970!

* feat: all nuon well-known runner facts follow same runner_* convention

we use well-known resource attributes w/ specific names as conventions
for facts about the objects doing the logging. `job.id` was the first
one we implemented and we omitted the `runner_*` prefix. in this pr, we
update it to use it across the board.