chore: switch runner to use otel-go (#6172)
* chore: get logs writing to api

* chore: remove unused deps

* chore: cleanup otlp client

* feat: introduce standard loggers across system

* fix: conflict follow ups

* remove

* chore: rename v2 to default org, update logic

* feat: add table sizes view + view directory

* fix: table sizes

* chore: add table view

* chore: update dependencies

* feat: wire up sandbox mode into jobs

* chore: more details on local sandbox mode

* fix: sandbox mode working

* fix: oapi spec not loading for runners

* addtl settings

* fix: more spec stuff

* fixes

* feat: wire up sandbox mode into runner

* fixes

* chore: add metrics

* fixes for configuration and more

* update

* chore: update sdk

* update

* update

* remove commenting

* update

* feat: pkg/gorm/clickhouse: skip migrations for default value -> null

in clickhouse, NOT NULL is not necessary. also, types have their own
default values. the clickhouse gorm package was not handlign this well
so we have added a special case to handle the case where gorm thinks the
a column needs to automigrate on the basis of the nullable status of a
column.

In effect, this does mean all of our tables can have rows with empty
columns or columns w/ the default values.

TODO: sort out the right gorm struct tag invocation to let the
automigrator know this is fine/intended.

* feat: ignore all default value -> null migrations

---------

Co-authored-by: fidiego <sayheytofred@gmail.com>