fix(services/ctl-api): stop routing traffic to dead pods (#7027)
* fix(services/ctl-api): stop routing traffic to dead pods

We think these annotations were set in the past, but were removed at
some point. Without, the target group used the default values of 15
second intervals, and two failed health checks. This caused the target
group to route traffic to dead pods for about 30 seconds.

This change will configure the target to do the following instead:

- Check health every 5 seconds.
- Mark a target as unhealthy after only 1 failed health check.

It will continue requiring 2 successful health checks before marking a
target healthy.

* fix(services/dashboard-ui): stop routing traffic to dead targets

* fix(services/wiki): stop routing traffic to dead targets

This is not a customer-facing service, and we don't have any alerting
hooked up to this. But, it's ALB was also missing the healthcheck
annotations, and we may as well keep them consistent.