feat: infra/eks: add configurable `additional_nodepools` (#7752)
we want a public nodepool in the prod and stage clusters but we don't
need them in the runner clusters. as a result, we need a way to add
additional nodepools in specific clusters only. this PR introduces a
variable in the `vars` files to support this. it takes a list of
nodepool configs.

at the time of writing we are only making a subset of the configs
configurable.

Co-authored-by: nuonbot <bot@nuon.co>