fix(infra/on-call): complete the pagerduty migration to terraform (#8742)
* fix(infra/on-call): complete the pagerduty migration to terraform

We set up Pagerduty manually a few months ago. We decided recently that
we wanted to start managing it with Terraform, and copied the manual
setup to TF, but didn't complete the migration.

This PR updates the Terraform config to use the correct service names
(removing the "-0" from the end.) This is required for events to route
services and create incidents. The old, manually-created services will
need to be deleted before applying this.

* fix(infra/on-call): escalate to engineering schedule

Incidents aren't being escalated to the Engineering rotation, so
whoever is on-call will not be paged.