feat: terraform workspaces MVP (#8845)
* feat(service/ctl-api): include sandbox and tf workspace in install

It looked like there was no way to get to a terraform workspace from an
install ID.  This PR includes the sandbox and it's terraform workspace
in the response from `GET /v1/installs/{install_id}`, so I can interact
with the sandbox's workspace from the install section of the UI.

I'm not sure this is the best way to do this, just wanted to unblock
myself.

I'm also wondering if it would make sense to add a `GET
/v1/installs/{install_id}/sandbox` endpoint. Currently, we use the app
sandbox data on the install sandbox page.

* feat(services/dashboard-ui): add terraform state to install sandbox page

Use updated API to display basic terraform workspace component on
install sandbox page.

- Display resources, outputs, and history
- Render Terraform backend config for use with the Terraform CLI

* feat: add terraform workspace feature toggle

* fix(services/dashboard-ui): fix tsc errors

* refactor(services/dashboard-ui): pr review changes