feat(bins/scripts): PRO-948 - nctl script to nuke all sandbox accounts (#8867)
* PRO-948 - nctl script to nuke all sandbox accounts

* added aws-nuke as a thing to install to our Brewfile
* added my own environment to our aws-nuke.yaml

The nuke-profile script is basically exactly the same thing as
`aws-nuke nuke -c "$CONFIG_FILE" --profile "$PROFILE" --no-alias-check --no-prompt --prompt-delay 5 $NO_DRY_RUN_FLAG`

The only reason for it's existence is that it seemed odd that there would be no single target version of the multi-verison.

The `nuke-sandbox-accounts` does similar logic as the above, across all the AWS profiles that are in your config that match the `^sandbox-` regex. (This excludes test and demo and such).  There is a further requirement that there must be a configuration against the account in the supplied aws-nuke.yaml

Maybe we can relax requirements around this, but I felt that starting very cautious was better than going crazy

* edited both nuke files to utilize std-error for error messages.

Updated aws-nuke.yaml, adding the retool-sandbox to the block list, and adding everyone else's sandboxes to the allow list.

---------

Co-authored-by: nuonbot <bot@nuon.co>