feat: Ja/missing tf state UI (#8912)
* fix(services/dashboard-ui): tf state empty message

It looks like Nat may have been in the middle of fixing my bug. The
`<Empty />` component was imported, but not used. Instead the component
just return `null`, so that the state compponent was entirely gone from
the page.

I added the `<Empty />` component with an appropriate message for the
user.

* feat(services/dashboard-ui): empty tf state still works

Replacing the entire terraform state component with the `<Empty />`
component had a couple of side effects:

- The section title was missing, which is confusing.
- The state dropdown was mising, so the user could no longer download a
  backend file. Being able to push to an empty workspace is something a
  user would use the backend file for.

This PR replaces only the contents of the state component with the
`<Empty />` component, instead of replacing it entirely.