feat(services/dashboard-ui): consistent design for empty components page (#8914)
Updated the empty state for the components page to be consistent with
the rest of the app. It now uses the `<Empty />` component, and links to
https://docs.nuon.co/concepts/components.

We don't have a custom empty graphic for components (like we do for
actions, for example,) so this just uses the table graphic. I think it's
good enough, but adding more empty graphics is a nice-to-have.

Co-authored-by: nuonbot <bot@nuon.co>