fix(bins/cli): dev command fails if any builds fail (#9003)
The dev command was printing that builds completed, even if one failed,
causing the workflow to continue even though it should stop.

There are two errors in the pollBuilds method, one for the entire group,
and one per build. The per build errors were overwriting the group
error. This PR fixes that.