fix: connect to github via cli in byoc installs (#9157)
The CLI uses our own github apps by default. You can configure it to use
a different app, just like you can configure it to use a different api,
but that's not a good UX.

This PR adds a `/connect-github` endpoint to the dashboard, and updates
the CLI to send users there. The dashboard will already be configured
with the correct github app for the byoc install the user is workin
with.