fix(bins/cli): login fails if it can't parse config file (#9176)
Viper is trying to parse the file extension of the config file, because
it wants to magically support multiple file types. It returns an error
if it can't recognize the extension, but in the `login` command we're
ignoring that.

This resulted in a silent failure to write to the config file for Retool
when logging in, because they're using config files with the format
`~/.nuon.nuon-dev-2`. Viper thinks `.nuon-dev-2` is the extension,
doesn't recognize it, and the newly fetched auth token isn't saved to
the config file.

This PR updates `login` to return an error, making it clear what
happened. I'm not sure if we can support Retool's preferred file naming
convention with Viper (or at least the version we're currently on.)