feat: [ctl-api] Create InstallConfig as part of Install creation (#9855)
This pathc updates the `CreateInstall` helper (used by ctl api to
create installs) to also optionally create an `InstallConfig` for
the `Install` in the same request. This allows us to have custom
options/config for an install before triggering any workflows.

This was needed as a way to set automatic approval for the
provsisioning workflow which gets automatically triggered after
and install is created.

Signed-off-by: Prem Saraswat <prem@nuon.co>