chore(ctl-api): return the approval response in its own response type (#10834)
chore(ctl-api): approval creation return type modification

modifies the workflow step approval creation endpoitn to
return a new, request-specific, type in order to prevent
naming collisions in generated SDKs

we've these two models:
  1. app.WorkflowStepApproval
  2. app.WorkflowStepApprovalResponse

Some endponts return a `WorkflowStepApproval` and some SDKs (python)
autogenerate a class for this response `WorkflowStepApprovalResponse`.
This collides with the class generated for the model
`WorkflowStepApprovalResponse`.

To avoid this, we return the whole approval instead of the response.

This was tested locally w/ the CLI and the dashboard.

Co-authored-by: nuonbot <bot@nuon.co>