fix: Skip executing Action Workflows that are not in app config (#10993)
This patch fixes a bug where action workflow runs were being
triggered for actions that have been removed from an app config.
This included action runs triggered as post/pre deploy hooks,
cron, syncing secrets, etc.

We now check if an action is part of the Install's current app
config, and skip executing it if that's not the case.

Signed-off-by: Prem Saraswat <prem@nuon.co>