chore(cli): small improvements (#11109)
* chore: add -w flag to installs workflows

* chore(cli): set app_id on install selection

a common flow is to select an install and then try to do work on it.
however, many downstream commands require and app_id to be configured,
but selecting the app is not a pre-requisite for selecting the install.

this change sets the app_id to the selected install's app id if the
values are not the same.