feat: CLI: Add `nuon installs toggle-sync` command (#11149)
* feat: CLI: Add `nuon installs toggle-sync` command

This patch adds a new command to cli, `nuon installs toggle-sync`.
This command toggles config file syncing for an install (enable it
if disabled, or vice-versa). For better ux in some cases, there are
also two mutually exclusive optional flags `--enable` and `--disable`
to explicitly enable or disable config file syncing for the install.

Signed-off-by: Prem Saraswat <prem@nuon.co>

* Update bins/cli/internal/services/installs/toggle_sync.go

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>

---------

Signed-off-by: Prem Saraswat <prem@nuon.co>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>
Co-authored-by: nuonbot <bot@nuon.co>