feat: install workflows: Wait for approval in a child workflow (#11252)
* feat: install workflows: Wait for approval in a child workflow

This PR patches the approvals awaiting logic to start a separate
child workflow. This allowed us to add support for using `ContinueAsNew`
while polling for approval, leading to cleaner history on the main
workflow, as well as getting rid of the workflow history size too
large errors with waiting for approvals for a long time.

Signed-off-by: Prem Saraswat <prem@nuon.co>

* use clearer names for variable

Signed-off-by: Prem Saraswat <prem@nuon.co>

* add some sanity checks

Signed-off-by: Prem Saraswat <prem@nuon.co>

* Fix poll deadline not being set correctly

* Only use the new approvals flow for workflows created after 26 Nov 2025

Signed-off-by: Prem Saraswat <prem@nuon.co>

* address comments around temporal timeout errors

Signed-off-by: Prem Saraswat <prem@nuon.co>

---------

Signed-off-by: Prem Saraswat <prem@nuon.co>
Co-authored-by: nuonbot <bot@nuon.co>